+++ babel.config.cjs
@@ -7,17 +7,17 @@ module.exports = {
   ],
   env: {
     cjs: {
+      browserslistEnv: "isomorphic-production",
       presets: [
         [
           '@babel/preset-env',
           {
             debug: false,
-            modules: false,
-            targets: {
-              node: '12.22.0',
-            },
+            modules: "commonjs",
+            loose: true,
+            useBuiltIns: false,
             forceAllTransforms: false,
-            ignoreBrowserslistConfig: true,
+            ignoreBrowserslistConfig: false,
             exclude: ['transform-function-name'],
           },
         ],
@@ -25,16 +25,13 @@ module.exports = {
       ],
       plugins: [
         [
-          '@babel/plugin-transform-modules-commonjs',
-          {
-            loose: true,
-          },
-        ],
-        [
           '@babel/plugin-transform-runtime',
           {
-            corejs: 3,
-            version: '^7',
+            corejs: { version: 3, proposals: false },
+            absoluteRuntime: false,
+            helpers: true,
+            regenerator: false,
+            version: '^7.22.15',
           },
         ],
         process.env.NODE_ENV !== 'test'
@@ -43,42 +40,15 @@ module.exports = {
       ].filter(Boolean),
     },
     es: {
+      browserslistEnv: "isomorphic-production",
       presets: [
         [
           '@babel/preset-env',
           {
             debug: false,
             modules: false,
-            targets: {
-              node: '12.22.0',
-            },
-            ignoreBrowserslistConfig: true,
-            exclude: ['transform-function-name'], // this is here because of https://github.com/babel/babel/discussions/12874
-          },
-        ],
-        '@babel/preset-typescript',
-      ],
-      plugins: [
-        '@babel/proposal-class-properties',
-        '@babel/proposal-object-rest-spread',
-        [
-          '@babel/plugin-transform-runtime',
-          {
-            corejs: 3,
-            version: '^7',
-            useESModules: true,
-          },
-        ],
-        [path.join(__dirname, './scripts/babel-plugin-add-import-extension.cjs'), { extension: 'js' }],
-      ],
-    },
-    browser: {
-      presets: [
-        [
-          '@babel/preset-env',
-          {
-            debug: false,
             useBuiltIns: false,
+            forceAllTransforms: false,
             ignoreBrowserslistConfig: false,
             exclude: ['transform-function-name'], // this is here because of https://github.com/babel/babel/discussions/12874
           },
@@ -86,17 +56,48 @@ module.exports = {
         '@babel/preset-typescript',
       ],
       plugins: [
-        '@babel/proposal-class-properties',
-        '@babel/proposal-object-rest-spread',
         [
           '@babel/plugin-transform-runtime',
           {
-            corejs: 3,
-            version: '^7',
+            corejs: { version: 3, proposals: false },
+            absoluteRuntime: false,
+            helpers: true,
+            regenerator: false,
             useESModules: true,
+            version: '^7.22.15',
           },
         ],
+        [path.join(__dirname, './scripts/babel-plugin-add-import-extension.cjs'), { extension: 'mjs' }],
       ],
+    },
+    browser: {
+      browserslistEnv: "browser-production",
+      presets: [
+        [
+          '@babel/preset-env',
+          {
+            debug: false,
+            modules: 'auto',
+            useBuiltIns: false,
+            forceAllTransforms: false,
+            ignoreBrowserslistConfig: false,
+            exclude: ['transform-function-name'], // this is here because of https://github.com/babel/babel/discussions/12874
+          },
+        ],
+        '@babel/preset-typescript',
+      ],
+      plugins: [
+        [
+          '@babel/plugin-transform-runtime',
+          {
+            corejs: { version: 3, proposals: false },
+            absoluteRuntime: false,
+            helpers: true,
+            regenerator: false,
+            version: '^7.22.15',
+          },
+        ],
+      ]
     },
   },
 };