+++ packages/apidom-core/src/clone/errors/DeepCloneError.ts
@@ -1,0 +1,5 @@
+import CloneError from './CloneError';
+
+class DeepCloneError extends CloneError {}
+
+export default DeepCloneError;