+++ packages/apidom-ns-asyncapi-2/src/refractor/plugins/replace-empty-element.ts
@@ -7,6 +7,7 @@ import {
   isStringElement,
   includesClasses,
   isArrayElement,
+  cloneDeep,
 } from '@swagger-api/apidom-core';
 
 import mediaTypes from '../../media-types';
@@ -1055,11 +1056,11 @@ const plugin = () => () => {
           elementFactory.call(
             { context: ancestor },
             undefined,
-            originalValue.meta.clone(),
-            originalValue.attributes.clone(),
+            cloneDeep(originalValue.meta),
+            cloneDeep(originalValue.attributes),
           ),
-          element.meta.clone(),
-          element.attributes.clone(),
+          cloneDeep(element.meta),
+          cloneDeep(element.attributes),
         );
       },
 
@@ -1080,8 +1081,8 @@ const plugin = () => () => {
         return elementFactory.call(
           { context: element },
           undefined,
-          element.meta.clone(),
-          element.attributes.clone(),
+          cloneDeep(element.meta),
+          cloneDeep(element.attributes),
         );
       },
     },