+++ packages/apidom-ns-json-schema-draft-4/src/refractor/visitors/generics/PatternedFieldsVisitor.ts
@@ -1,7 +1,14 @@
 import stampit from 'stampit';
 import { F as stubFalse } from 'ramda';
 import { noop } from 'ramda-adjunct';
-import { ObjectElement, Element, MemberElement, BREAK } from '@swagger-api/apidom-core';
+import {
+  ObjectElement,
+  Element,
+  MemberElement,
+  BREAK,
+  cloneDeep,
+  toValue,
+} from '@swagger-api/apidom-core';
 
 import SpecificationVisitor from '../SpecificationVisitor';
 
@@ -25,17 +32,17 @@ const PatternedFieldsJsonObjectVisitor = stampit(SpecificationVisitor, {
       // @ts-ignore
       objectElement.forEach((value: Element, key: Element, memberElement: MemberElement) => {
         if (
-          !this.ignoredFields.includes(key.toValue()) &&
-          this.fieldPatternPredicate(key.toValue())
+          !this.ignoredFields.includes(toValue(key)) &&
+          this.fieldPatternPredicate(toValue(key))
         ) {
           const specPath = this.specPath(value);
           const patternedFieldElement = this.toRefractedElement(specPath, value);
-          const newMemberElement = new MemberElement(key.clone(), patternedFieldElement);
+          const newMemberElement = new MemberElement(cloneDeep(key), patternedFieldElement);
           this.copyMetaAndAttributes(memberElement, newMemberElement);
           newMemberElement.classes.push('patterned-field');
           this.element.content.push(newMemberElement);
-        } else if (!this.ignoredFields.includes(key.toValue())) {
-          this.element.content.push(memberElement.clone());
+        } else if (!this.ignoredFields.includes(toValue(key))) {
+          this.element.content.push(cloneDeep(memberElement));
         }
       });