+++ packages/apidom-ns-openapi-3-1/src/refractor/visitors/open-api-3-1/schema/$vocabularyVisitor.ts
@@ -1,11 +1,11 @@
 import stampit from 'stampit';
-import { ObjectElement, BREAK } from '@swagger-api/apidom-core';
+import { ObjectElement, BREAK, cloneDeep } from '@swagger-api/apidom-core';
 import { FallbackVisitor } from '@swagger-api/apidom-ns-openapi-3-0';
 
 const $vocabularyVisitor = stampit(FallbackVisitor, {
   methods: {
     ObjectElement(objectElement: ObjectElement) {
-      this.element = objectElement.clone();
+      this.element = cloneDeep(objectElement);
       this.element.classes.push('json-schema-$vocabulary');
 
       return BREAK;