+++ packages/apidom-ns-openapi-3-1/src/refractor/visitors/open-api-3-1/schema/TypeVisitor.ts
@@ -1,17 +1,17 @@
 import stampit from 'stampit';
-import { StringElement, ArrayElement, BREAK } from '@swagger-api/apidom-core';
+import { StringElement, ArrayElement, BREAK, cloneDeep } from '@swagger-api/apidom-core';
 import { FallbackVisitor } from '@swagger-api/apidom-ns-openapi-3-0';
 
 const TypeVisitor = stampit(FallbackVisitor, {
   methods: {
     StringElement(stringElement: StringElement) {
-      this.element = stringElement.clone();
+      this.element = cloneDeep(stringElement);
       this.element.classes.push('json-schema-type');
 
       return BREAK;
     },
     ArrayElement(arrayElement: ArrayElement) {
-      this.element = arrayElement.clone();
+      this.element = cloneDeep(arrayElement);
       this.element.classes.push('json-schema-type');
 
       return BREAK;