+++ packages/apidom-error/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.76.2](https://github.com/swagger-api/apidom/compare/v0.76.1...v0.76.2) (2023-09-08)
+
+### Bug Fixes
+
+- **build:** apply polyfills with transform runtime babel plugin ([#3132](https://github.com/swagger-api/apidom/issues/3132)) ([4b4754d](https://github.com/swagger-api/apidom/commit/4b4754da33a4a25d47f926c8a14b59cfdfa6b50c))
+
+## [0.76.1](https://github.com/swagger-api/apidom/compare/v0.76.0...v0.76.1) (2023-09-07)
+
+### Bug Fixes
+
+- **error:** polyfill AggregateError in older environments ([#3129](https://github.com/swagger-api/apidom/issues/3129)) ([484d9e6](https://github.com/swagger-api/apidom/commit/484d9e6c34534aa37c696a5154af72cfcc9de09c)), closes [#3128](https://github.com/swagger-api/apidom/issues/3128)
+
 # [0.76.0](https://github.com/swagger-api/apidom/compare/v0.75.0...v0.76.0) (2023-09-01)
 
 ### Features